pythonmodel/requirements.txt
python-dotenv
transformers
pika
